workshop/f1-leaderboard-workshop/backend/requirements.txt
# Core web framework and server
fastapi>=0.104.1,<0.116.0
uvicorn[standard]>=0.24.0,<0.32.0

# Kafka integration
confluent-kafka[schema-registry]>=2.11.1,<3.0.0

# Data processing and validation
pydantic>=2.5.0,<3.0.0
pyyaml>=6.0.1,<7.0.0
python-multipart>=0.0.6,<1.0.0

# Schema validation
referencing>=0.37.0,<1.0.0
jsonschema>=4.25.1,<5.0.0
fastavro>=1.12.1,<2.0.0
orjson>=3.9.0,<4.0.0